requirements.txt
boto3
datasets
huggingface_hub
tqdm
pyyaml